feat(Gyoungmin-Min):[BOJ, 1966] 프린터큐 https://www.acmicpc.net/problem/1966
